RATIS-470. De-couple the LogService "index" from the RaftLog "index"
We've been using the Raft log index as the "record offset" that we expose
to users via the LogService API. However, because one message that we push
to the Raft log may contain many LogService records, this generates
incorrect results.

We have to "hide" the Raft log index, and identify just the LogService records
so that we're exposing only their Log's information (e.g. none of the interal
Raft quorum configuration entries).

Closes #7