RATIS-473. Clarify size and length methods on LogStream
getLength() was meant to be the number of records in a log
but was implemented as the size in bytes of the records in
the log. This commit adds a getSize() which takes the old
implementation of getLength(), and then reimplements
getLength() as the number of records in the log.

Signed-off-by: Vladimir Rodionov <vrodionov@apache.org>
Signed-off-by: Rajeshbabu Chintaguntla <rajeshbabu@apache.org>